+++ sdk/highlight-cloudflare/tsup.config.ts
@@ -5,4 +5,5 @@ export default defineConfig({
 	format: ['cjs', 'esm'],
 	dts: true,
 	sourcemap: true,
+	treeshake: 'smallest',
 })