+++ sdk/highlight-next/src/util/with-highlight-edge.ts
@@ -1,6 +1,5 @@
 import type { NodeOptions } from '@highlight-run/node'
 import { H } from './highlight-edge'
-import { ExtendedExecutionContext } from './types'
 
 export type NextFetchEvent = {
 	params: Promise<Record<string, string>>
@@ -16,42 +15,26 @@ export type EdgeHandler = (
 
 export function Highlight(env: HighlightEnv) {
 	return function withHighlight(handler: EdgeHandler) {
-		return async function (request: NextRequest, context: NextFetchEvent) {
-			if (env.enableFsInstrumentation) {
-				console.warn(
-					'enableFsInstrumentation is incompatible with Edge... disabling now.',
-				)
-
-				env.enableFsInstrumentation = false
-			}
-
-			H.initEdge(
-				request,
-				env,
-				context as unknown as ExtendedExecutionContext,
+		if (env.enableFsInstrumentation) {
+			console.warn(
+				'enableFsInstrumentation is incompatible with Edge... disabling now.',
 			)
 
+			env.enableFsInstrumentation = false
+		}
+		H.initEdge(env)
+		return async function (request: NextRequest, context: NextFetchEvent) {
 			try {
-				const response = await H.runWithHeaders(
+				return await H.runWithHeaders(
 					`${request.method?.toUpperCase()} - ${request.url}`,
-					request.headers as any,
-					async () => {
+					request.headers,
+					async (span) => {
+						span.setAttribute('next.runtime', 'edge')
 						return await handler(request, context)
 					},
 				)
-
-				H.sendResponse(response)
-
-				return response
-			} catch (error) {
-				const { secureSessionId, requestId } = H.parseHeaders(
-					request.headers as any,
-				)
-				if (error instanceof Error) {
-					H.consumeError(error, secureSessionId, requestId)
-				}
-
-				throw error
+			} finally {
+				await H.flush()
 			}
 		}
 	}