+++ sdk/highlight-remix/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @highlight-run/remix
 
+## 2.0.84
+
+### Patch Changes
+
+-   Updated dependencies [31dc610]
+-   Updated dependencies [31dc610]
+-   Updated dependencies [31dc610]
+-   Updated dependencies [31dc610]
+    -   @highlight-run/node@3.12.0
+    -   @highlight-run/react@12.1.0
+
+## 2.0.83
+
+### Patch Changes
+
+-   Updated dependencies [2b8cba8]
+-   Updated dependencies [baed388]
+    -   highlight.run@9.11.1
+    -   @highlight-run/node@3.11.1
+    -   @highlight-run/react@12.0.1
+
 ## 2.0.82
 
 ### Patch Changes